+++ extensions/browser-bookmarks/src/hooks/useAvailableBrowsers.ts
@@ -9,6 +9,7 @@ export const BROWSERS_BUNDLE_ID = {
   chrome: "com.google.chrome",
   chromeBeta: "com.google.chrome.beta",
   chromeDev: "com.google.chrome.dev",
+  dia: "company.thebrowser.dia",
   firefox: "org.mozilla.firefox",
   firefoxDev: "org.mozilla.firefoxdeveloperedition",
   island: "io.island.island",